+++ libraries/evaluators/tcc/Evaluator.tcc
@@ -65,7 +65,7 @@ namespace evaluators
         if (v.size() == 0) return;
 
         os << v[0];
-        for (uint64_t j = 1; j < v.size(); ++j)
+        for (size_t j = 1; j < v.size(); ++j)
         {
             os << '\t' << v[j];
         }
